requirements.txt
fastapi==0.109.0
uvicorn==0.26.0
pillow==10.2.0
